backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
chromadb
sentence-transformers==3.1.1
pydantic==2.9.1
python-multipart==0.0.12
rank-bm25==0.2.2
pymongo[srv]
requests
openai
python-dotenv
